+++ src/ui-history.c
@@ -64,8 +64,8 @@ void history_display(void)
 		row = 0;
 		for (i = first_item; row <= page_size && i < max_item; i++)
 		{
-			strnfmt(buf, sizeof(buf), "%10d%7d\'  %s",
-				history_list_local[i].turn,
+			strnfmt(buf, sizeof(buf), "%10ld%7d\'  %s",
+				(long)history_list_local[i].turn,
 				history_list_local[i].dlev * 50,
 				history_list_local[i].event);
 
@@ -136,8 +136,8 @@ void dump_history(ang_file *file)
 	file_putf(file, "      Turn   Depth  Note\n");
 
 	for (i = 0; i < max_item; i++) {
-		strnfmt(buf, sizeof(buf), "%10d%7d\'  %s",
-				history_list_local[i].turn,
+		strnfmt(buf, sizeof(buf), "%10ld%7d\'  %s",
+				(long)history_list_local[i].turn,
 				history_list_local[i].dlev * 50,
 				history_list_local[i].event);
