+++ src/wiz-spoil.c
@@ -393,7 +393,7 @@ void spoil_artifact(const char *fname)
 	/* Dump the header */
 	spoiler_underline(format("Artifact Spoilers for %s", buildid), '=');
 
-	text_out("\n Randart seed is %u\n", seed_randart);
+	text_out("\n Randart seed is %lu\n", (unsigned long)seed_randart);
 
 	/* List the artifacts by tval */
 	for (i = 0; group_artifact[i].tval; i++) {
@@ -450,9 +450,9 @@ void spoil_artifact(const char *fname)
 			 * artifact can appear, its rarity, its weight, and
 			 * its power rating.
 			 */
-			text_out("\nMin Level %u, Max Level %u, Generation chance %u, Power %d, %d.%d lbs\n",
+			text_out("\nMin Level %u, Max Level %u, Generation chance %u, Power %ld, %d.%d lbs\n",
 					 art->alloc_min, art->alloc_max, art->alloc_prob,
-					 object_power(obj, false, NULL), (art->weight / 10),
+					 (long)object_power(obj, false, NULL), (art->weight / 10),
 					 (art->weight % 10));
 
 			if (OPT(player, birth_randarts)) text_out("%s.\n", art->text);
